Split wrappers into their own files, added N-HiTS
